+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to Sample Commander will be documented in this file.
 
 ## [2.0.1] - yet to be released
 
+### Added
+* Files can now be selected by drawing a selection box around them by mouse (as in Windows Explorer, including the inversion behavior when holding the `Ctrl` key while drawing).
+
 ### Changed
 * minor internal refactorings